+++ smx/vm.py
@@ -1,17 +1,37 @@
 from __future__ import annotations
 
+import ctypes
+import dataclasses
 import logging
 import struct
 import sys
-from ctypes import *
+import typing
+from copy import deepcopy
+from ctypes import (
+    addressof,
+    c_byte,
+    c_char_p,
+    c_float,
+    c_int16,
+    c_int8,
+    c_void_p,
+    memmove,
+    memset,
+    POINTER,
+    pointer,
+    sizeof,
+)
 from datetime import datetime
+from enum import IntFlag
 from pathlib import Path
-from typing import Any, Dict, List, Tuple, Type, TYPE_CHECKING, TypeVar
+from typing import Any, Dict, List, NamedTuple, Tuple, Type, TYPE_CHECKING, TypeVar, Union
 
-from smx.definitions import cell, PyCSimpleType, ucell
+from smx.definitions import cell, PyCSimpleType, RTTIControlByte, SP_MAX_EXEC_PARAMS, ucell
 from smx.exceptions import SourcePawnPluginError, SourcePawnPluginNativeError
 from smx.opcodes import opcodes, SourcePawnInstruction
 from smx.pawn import SMXInstructions
+from smx.rtti import RTTI
+from smx.sourcemod.natives.base import convert_return_value, WritableString
 from smx.sourcemod.system import SourceModSystem
 
 if TYPE_CHECKING:
@@ -101,6 +121,7 @@ class SourcePawnAbstractMachine:
         # XXX(zk): is this necessary? does ctypes automatically zero out requested memory?
         memset(addressof(self.heap) + self.plugin.datasize, 0, self.plugin.memsize - self.plugin.datasize)
 
+        self.HEA = self.plugin.datasize
         self.STP = len(self.heap)
         self.STK = self.STP
         self.FRM = self.STK  # XXX: this is a guess. Is this correct?
@@ -144,8 +165,8 @@ class SourcePawnAbstractMachine:
         return val
 
     def _getheap(self, offset: int, ctype: Type[CType]) -> Any:
-        heap = cast(self.heap, POINTER(ctype))
-        heap_ptr = cast(pointer(heap), POINTER(c_void_p))
+        heap = ctypes.cast(self.heap, POINTER(ctype))
+        heap_ptr = ctypes.cast(pointer(heap), POINTER(c_void_p))
         heap_ptr.contents.value += offset
         return heap.contents.value
 
@@ -178,7 +199,7 @@ class SourcePawnAbstractMachine:
 
     def _sp_ctof(self, val: cell) -> float:
         """Cast a cell to a float"""
-        return cast(pointer(val), POINTER(c_float)).contents.value
+        return ctypes.cast(pointer(val), POINTER(c_float)).contents.value
 
     def _sp_ftoc(self, val: float) -> cell:
         """Shove a float into a cell"""
@@ -270,8 +291,8 @@ class SourcePawnAbstractMachine:
             raise NotImplementedError('Native %s not implemented, yet' %
                                       native.name)
 
-        params = cast(self.heap, POINTER(cell))
-        params_ptr = cast(pointer(params), POINTER(c_void_p))
+        params = ctypes.cast(self.heap, POINTER(cell))
+        params_ptr = ctypes.cast(pointer(params), POINTER(c_void_p))
         params_ptr.contents.value += paramoffs
 
         return pyfunc(params)
@@ -349,40 +370,322 @@ class SourcePawnAbstractMachine:
                     spew_line = f'{spew_line:<70}-> {rval}'
 
                 if self.runtime.spew_stack:
-                    stk = [f'[{hex(addr)}] {hex(value.value)}' for addr, value in self._stack]
+                    stk = [f'[{hex(addr)}/{hex(self.STK-addr)}] {hex(value.value)}' for addr, value in self._stack]
                     spew_line = f'{spew_line:<100} STK: {", ".join(stk)}'
 
                 print(spew_line)
 
 
-class PluginFunction:
-    def __init__(self, runtime: SourcePawnPluginRuntime, func_id, code_offs):
+class ParamCopyFlag(IntFlag):
+    COPYBACK = (1 << 0)
+
+
+class ParamStringFlag(IntFlag):
+    UTF8 = (1 << 0)
+    COPY = (1 << 1)
+    BINARY = (1 << 2)
+
+
+ParamScalarValueT = Union[int, float]
+ParamArrayValueT = Union[str, bytes, List[ParamScalarValueT]]
+ParamValueT = Union[ParamScalarValueT, ParamArrayValueT]
+
+
+@dataclasses.dataclass
+class ParamInfo:
+    value: ParamValueT | None = None
+
+    #: Whether the input was a scalar or a list of values
+    is_scalar: bool = True
+
+    #: Copy-back flags
+    flags: ParamCopyFlag = 0
+
+    # XXX(zk): this is a terrible name. ported from SM
+    #: Whether this is marked as being used
+    marked: bool = False
+
+    #: Size of array in cells, or string in bytes
+    size: int = 0
+
+    #: Whether this is a string
+    is_string: bool = False
+
+    #: String flags
+    string_flags: ParamStringFlag = 0
+
+    #: Local addr of allocated memory for the param
+    local_addr: int = -1
+    #: Physical addr of allocated memory for the param
+    phys_addr: int = -1
+
+
+class CallableReturnValue(NamedTuple):
+    rval: Any
+    args: List[ParamValueT]
+
+
+class ICallable:
+    def push_cell(self, value: int):
+        raise NotImplementedError
+
+    def push_cell_by_ref(self, value: int, flags: ParamCopyFlag = 0):
+        raise NotImplementedError
+
+    def push_float(self, value: float):
+        raise NotImplementedError
+
+    def push_float_by_ref(self, value: float, flags: ParamCopyFlag = 0):
+        raise NotImplementedError
+
+    # TODO(zk): allow ctypes inputs?
+    def push_array(self, value: List[int | float] | None, flags: ParamCopyFlag = 0):
+        raise NotImplementedError
+
+    def push_string(self, value: str):
+        raise NotImplementedError
+
+    def push_string_ex(self, value: str, length: int, string_flags: ParamStringFlag = 0, copy_flags: ParamCopyFlag = 0):
+        raise NotImplementedError
+
+    # TODO(zk): return result
+    def execute(self) -> Tuple[CallableReturnValue | None, int]:
+        raise NotImplementedError
+
+    # TODO(zk): return result
+    def invoke(self) -> Tuple[CallableReturnValue | None, bool]:
+        raise NotImplementedError
+
+
+def _new_callable_params() -> List[ParamInfo]:
+    return [ParamInfo() for _ in range(SP_MAX_EXEC_PARAMS)]
+
+
+class PluginFunction(ICallable):
+    def __init__(self, runtime: SourcePawnPluginRuntime, func_id: int, code_offs: int):
         self.runtime: SourcePawnPluginRuntime = runtime
         self.func_id = func_id
         self.code_offs = code_offs
 
+        self.rtti_method = self.runtime.plugin.rtti_methods_by_addr[code_offs]
+        self.rtti_args = self.rtti_method.rtti.args
+
+        #: Holds arguments for the current invocation
+        self._params: List[ParamInfo] = []
+
     def __call__(self, *args):
+        call_rval, did_succeed = self.call(*args)
+        return call_rval.rval
+
+    def call(self, *args):
         if not self.runtime.amx.initialized:
             self.runtime.amx.init()
 
-        # Let's not clobber the current CIP, if possible
+        self._params = []
+
+        for i, arg in enumerate(args):
+            if i >= len(self.rtti_args):
+                if self.rtti_args[-1].is_variadic:
+                    rtti_arg = self.rtti_args[-1]
+                else:
+                    raise ValueError(f'Expected {len(self.rtti_args)} arguments, got {len(args)}')
+            else:
+                rtti_arg = self.rtti_args[i]
+
+            if isinstance(arg, (int, bool, float)):
+                if rtti_arg.is_by_ref:
+                    # XXX(zk): are there any cases we *don't* want to copy back?
+                    self.push_cell_by_ref(arg, ParamCopyFlag.COPYBACK)
+                else:
+                    self.push_cell(arg)
+
+            elif isinstance(arg, (str, bytes)):
+                if rtti_arg.is_by_ref:
+                    string_flags = ParamStringFlag.UTF8 if isinstance(arg, str) else ParamStringFlag.BINARY
+                    string_flags |= ParamStringFlag.COPY
+                    self.push_string_ex(arg, len(arg), string_flags, ParamCopyFlag.COPYBACK)
+                else:
+                    self.push_string(arg)
+
+            else:
+                try:
+                    arg_iter = iter(arg)
+                except TypeError:
+                    raise TypeError(f'Invalid argument type {type(arg)}: {arg!r}')
+
+                arg = list(arg_iter)
+                if rtti_arg.is_by_ref:
+                    self.push_array(arg, ParamCopyFlag.COPYBACK)
+                else:
+                    self.push_array(arg)
+
+        return self.invoke()
+
+    def _add_param(self) -> ParamInfo:
+        if len(self._params) >= SP_MAX_EXEC_PARAMS:
+            raise ValueError(f'Cannot add more than {SP_MAX_EXEC_PARAMS} parameters')
+
+        param = ParamInfo()
+        self._params.append(param)
+        return param
+
+    def push_cell(self, value: int):
+        param = self._add_param()
+        param.value = value
+        param.marked = False
+        return param
+
+    def push_cell_by_ref(self, value: int, flags: ParamCopyFlag = 0):
+        param = self.push_array([value], flags)
+        param.is_scalar = True
+        return param
+
+    def push_float(self, value: float):
+        return self.push_cell(typing.cast(int, value))
+
+    def push_float_by_ref(self, value: float, flags: ParamCopyFlag = 0):
+        return self.push_cell_by_ref(typing.cast(int, value), flags)
+
+    def push_array(self, value: List[int | float] | None, flags: ParamCopyFlag = 0):
+        param = self._add_param()
+        param.value = value
+        param.is_scalar = False
+        param.marked = True
+        param.size = len(value) if value is not None else 0
+        param.flags = flags if value is not None else 0
+        param.is_string = False
+        return param
+
+    def push_string(self, value: str):
+        return self.push_string_ex(value, len(value) + 1, ParamStringFlag.COPY, 0)
+
+    def push_string_ex(self, value: str, length: int, string_flags: ParamStringFlag = 0, copy_flags: ParamCopyFlag = 0):
+        param = self._add_param()
+        param.value = value
+        param.is_scalar = False
+        param.marked = True
+        param.size = length
+        param.flags = copy_flags
+        param.is_string = True
+        param.string_flags = string_flags
+        return param
+
+    def execute(self) -> Tuple[CallableReturnValue | None, int]:
+        # TODO(zk): clear pending exceptions
+
+        rval, err = self.invoke()
+        if not err:
+            # TODO(zk): return exception code
+            return None, -1
+
+        # TODO(zk): constant for "no exception"
+        return rval, 0
+
+    def invoke(self) -> Tuple[CallableReturnValue | None, bool]:
+        # TODO(zk): check runnable
+        # TODO(zk): check error state
+
+        # Copy params, allowing reentrancy (calls within calls, yo)
+        params = deepcopy(self._params)
+
+        args: List[int] = []
+        for param in params:
+            # Is this marked as an array?
+            if param.marked:
+                if not param.is_string:
+                    # Allocate a normal/generic array
+                    param.local_addr, param.phys_addr = self.runtime.heap_alloc(param.size)
+                    if param.value is not None:
+                        values = [convert_return_value(v) for v in param.value]
+                        array = (cell * len(values))(*values)
+                        self.runtime.amx._writeheap(param.local_addr, array)
+
+                else:  # is_string
+                    num_cells = (param.size + sizeof(cell) - 1) // sizeof(cell)
+                    param.local_addr, param.phys_addr = self.runtime.heap_alloc(num_cells)
+
+                    if param.string_flags & ParamStringFlag.COPY and param.value is not None:
+                        # TODO(zk): forego these flags for PluginFunction
+                        if param.string_flags & ParamStringFlag.UTF8:
+                            assert isinstance(param.value, str)
+                            value = param.value.encode('utf-8')
+                        elif param.string_flags & ParamStringFlag.BINARY:
+                            assert isinstance(param.value, bytes)
+                            value = param.value
+                        else:
+                            assert isinstance(param.value, str)
+                            value = param.value.encode('latin-1')
+                        buf = WritableString(self.runtime.amx, param.local_addr, param.size)
+                        buf.write(value)
+
+                args.append(param.local_addr)
+
+            else:  # not array
+                args.append(convert_return_value(param.value))
+
+        # TODO(zk): handle exception states
+        rval = self._call(args)
+
+        out_args_rev: List[ParamValueT | None] = []
+        for i, param in reversed(tuple(enumerate(params))):
+            if not param.marked:
+                out_args_rev.append(param.value)
+                continue
+
+            if i < len(self.rtti_args):
+                rtti_arg = self.rtti_args[i]
+            elif self.rtti_args[-1].is_variadic:
+                rtti_arg = self.rtti_args[-1]
+            else:
+                # Default cell type
+                rtti_arg = RTTI(self.runtime.plugin, type=RTTIControlByte.ANY)
+
+            if param.flags & ParamCopyFlag.COPYBACK:
+                if param.is_string:
+                    out_args_rev.append(rtti_arg.interpret_value(param.local_addr, self.runtime.amx))
+                elif param.is_scalar:
+                    val = self.runtime.amx._getheapcell(param.local_addr)
+                    out_args_rev.append(rtti_arg.interpret_value(val, self.runtime.amx))
+                else:
+                    cells = self.runtime.amx._getheap(param.local_addr, (cell * param.size))
+                    values = [rtti_arg.interpret_value(c.value, self.runtime.amx) for c in cells]
+                    out_args_rev.append(values)
+
+            self.runtime.heap_pop(param.local_addr)
+
+        call_rval = CallableReturnValue(rval, out_args_rev[::-1])
+        return call_rval, True
+
+    def _call(self, args: List[int]) -> Any:
+        if not self.runtime.amx.initialized:
+            self.runtime.amx.init()
+
+        # Save current runtime state
+        prev_stk = self.runtime.amx.STK
+        prev_hea = self.runtime.amx.HEA
+        prev_frm = self.runtime.amx.FRM
+        prev_hp_scope = self.runtime.amx._hp_scope
         prev_cip = self.runtime.amx.CIP
 
         # CODE 0 always seems to be a HALT instruction,
         # so when we RETN, we'll halt.
         self.runtime.amx.CIP = 0
 
-        # TODO: handle args
-        num_args = 0
-        self.runtime.amx._push(num_args)
+        self.runtime.amx._push(len(args))
+        for arg in reversed(args):
+            self.runtime.amx._push(arg)
 
-        # Enter the function, taking care to handle frame state
+        # Abide by calling conventions
         self.runtime.amx.instructions.call(self.runtime.amx, self.code_offs)
 
         # Execute the function body
         rval = self.runtime.amx._execute(self.code_offs)
 
-        # Restore the previous CIP
+        # Restore the previous runtime state
+        self.runtime.amx.STK = prev_stk
+        self.runtime.amx.HEA = prev_hea
+        self.runtime.amx.FRM = prev_frm
+        self.runtime.amx._hp_scope = prev_hp_scope
         self.runtime.amx.CIP = prev_cip
 
         return rval
@@ -478,3 +781,47 @@ class SourcePawnPluginRuntime:
 
         self.amx.smsys.timers.poll_for_timers()
         return rval
+
+    def heap_alloc(self, num_cells: int) -> Tuple[int, int]:
+        """Allocates a bounded block of memory on the secondary stack of a plugin
+
+        Note that although called a heap, it is in fact a stack.
+
+        :param num_cells:
+            Number of cells to allocate
+
+        :return:
+            A tuple of (local_addr, phys_addr), where local_addr is the heap offset
+            used to deallocate the memory, and phys_addr is the physical address
+            where values may be written to
+
+        """
+        num_bytes = num_cells * sizeof(cell)
+
+        # TODO(zk): check stack margin
+
+        bounds_addr = self.amx.HEA
+        self.amx._writeheap(bounds_addr, cell(num_cells))
+        self.amx.HEA += sizeof(cell)
+        local_addr = self.amx.HEA
+        phys_addr = addressof(self.amx.heap) + self.amx.HEA
+
+        self.amx.HEA += num_bytes
+
+        return local_addr, phys_addr
+
+    def heap_pop(self, local_addr: int) -> None:
+        """Pops a heap address off the heap/secondary stack
+
+        Use this to free memory allocated with heap_alloc()
+
+        Note that in SourcePawn, the heap is in fact a bottom-up stack.
+        Deallocations with this method should be performed in precisely the REVERSE order.
+        """
+        bounds_addr = local_addr - sizeof(cell)
+        num_bytes = self.amx._getheapcell(bounds_addr) * sizeof(cell)
+        if self.amx.HEA - num_bytes != local_addr:
+            # TODO(zk): richer exception, with SM's code (5, SP_ERROR_INVALID_ADDRESS)
+            raise RuntimeError('Invalid address')
+
+        self.amx.HEA = bounds_addr